nfc-inventory-server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pymysql
google-auth
